% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{create_url}
\alias{create_url}
\title{Create URL from a unique identifier}
\usage{
create_url(x, user = NULL, type = NULL)
}
\arguments{
\item{x}{Unique identifier (e.g. ID column from \code{erowid})}

\item{user}{Dreamjournal username}

\item{type}{"erowid" or "dreamjournal"}
}
\value{
A character string containing the full URL.
}
\description{
To avoid repetition, the raw data frames don't contain reports' URLs. This
function constructs a URL given an Erowid or Dreamjournal identifier.
}
\examples{
# To obtain an Erowid URL
create_url(erowid$id[1])
# To obtain a dreamjournal URL
create_url(dreamjournal$id[1], dreamjournal$user[1])

}
\author{
Matti Vuorre \email{mv2521@columbia.edu}
}

