% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dreamjournal}
\alias{dreamjournal}
\title{Dreamjournal Dream Reports}
\format{A data frame (tibble) with 131661 rows and 15 variables:
\describe{
  \item{id}{Report ID}
  \item{dream}{Name of the dream, or title}
  \item{user}{Author's username}
  \item{date}{Date when report was logged}
  \item{rating}{Rating}
  \item{cohesion}{Cohesion of dream}
  \item{lucidity}{Rating of the dream's lucidity}
  \item{views}{Number of views}
  \item{themes}{List-column}
  \item{settings}{List-column}
  \item{characters}{List-column}
  \item{emotions}{List-column}
  \item{activities}{List-column}
  \item{keywords}{List-column}
  \item{text}{Text body of the dream report}
}}
\source{
\url{https://www.erowid.org/experiences/exp_front.shtml}
}
\usage{
head(dreamjournal)  # Print first 6 rows
}
\description{
Tidy data frame ready for text analysis.
\url{http://www.dreamjournal.net/}).
}
\keyword{datasets}

