% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{erowid}
\alias{erowid}
\title{Erowid Experience Reports}
\format{A data frame (tibble) with 24788 rows and 14 variables:
\describe{
  \item{id}{Report ID on Erowid database}
  \item{title}{Title of the report}
  \item{author}{Name or pseudonym of author}
  \item{gender}{Author's gender}
  \item{age}{Age of author at time of experience}
  \item{kg}{weight of author in kg}
  \item{year}{Year when experience occurred}
  \item{published}{When report was published on Erowid}
  \item{rating}{Categorical rating of report quality}
  \item{views}{Number of views}
  \item{substance}{The (list of) substances reported}
  \item{text}{Report text}
  \item{erowid_warning}{Logical: Did report contain Erowid warning}
  \item{dosechart}{List-column: Extra info on substance and dose}
}}
\source{
\url{https://www.erowid.org/experiences/exp_front.shtml}
}
\usage{
head(erowid)  # Print first 6 rows
}
\description{
Tidy data frame ready for text analysis.
\url{https://www.erowid.org/experiences/exp_front.shtml}).
}
\keyword{datasets}

